requirements.txt
s3fs
boto3
pandas
awslambdaric
snowflake-connector-python